+++ app/controllers/concerns/authenticable.rb
@@ -1,0 +1,26 @@
+module Authenticable
+  extend ActiveSupport::Concern
+
+  def authenticate_user_using_x_auth_token
+    user_email = request.headers["X-Auth-Email"]
+    auth_token = request.headers["X-Auth-Token"].presence
+    user = user_email && User.find_by_email(user_email)
+
+    if user && auth_token &&
+      ActiveSupport::SecurityUtils.secure_compare(
+        user.authentication_token, auth_token
+      )
+      @current_user = user
+    else
+      render status: :unauthorized, json: {
+        errors: [t('session.could_not_auth')] 
+      }
+    end
+  end
+
+  private
+
+  def current_user
+    @current_user
+  end
+end
