+++ app/controllers/polls_controller.rb
@@ -1,7 +1,9 @@
 class PollsController < ApplicationController
+  after_action :verify_authorized, only: %i[update destroy]
   before_action :authenticate_user_using_x_auth_token, except: :index
   before_action :load_poll, only: %i[show update destroy]
   before_action :load_options, :load_responses, only: %i[show]
+  before_action :authorize_poll, only: %i[update destroy]
 
   def index
     polls = Poll.all.order('created_at DESC')
@@ -12,7 +14,9 @@ class PollsController < ApplicationController
     @poll = Poll.new(poll_params)
 
     if @poll.save
-      render status: :ok, json: { notice: t('successfully_created', entity: 'Poll') }
+      render status: :ok, json: {
+        notice: t('successfully_created', entity: 'Poll')
+      }
     else
       errors = @poll.errors.full_messages
       render status: :unprocessable_entity, json: { errors: errors }
@@ -26,9 +30,10 @@ class PollsController < ApplicationController
   end
 
   def update
-    authorize @poll
     if @poll.update(poll_params)
-      render status: :ok, json: { notice: 'Successfully updated poll.' }
+      render status: :ok, json: {
+        notice: t('successfully_updated', entity: 'Poll')
+      }
     else
       errors = @poll.errors.full_messages
       render status: :unprocessable_entity, json: { errors: errors }
@@ -36,9 +41,10 @@ class PollsController < ApplicationController
   end
 
   def destroy
-    authorize @poll
     if @poll.destroy
-      render status: :ok, json: { notice: 'Successfully deleted poll.' }
+      render status: :ok, json: { 
+        notice: t('successfully_deleted', entity: 'Poll')
+      }
     else
       errors = @poll.errors.full_messages
       render status: :unprocessable_entity, json: { errors: errors }
@@ -69,5 +75,9 @@ class PollsController < ApplicationController
     @responses = Response.where(poll_id: params[:id])
     rescue ActiveRecord::RecordNotFound => errors
       render json: {errors: errors}
+  end
+
+  def authorize_poll
+    authorize @poll
   end
 end
