+++ app/controllers/users_controller.rb
@@ -1,10 +1,4 @@
 class UsersController < ApplicationController
-  # before_action :authenticate_user_using_x_auth_token, only: [:index]
-
-  def index
-    users = User.all.as_json(only: %i[id first_name last_name])
-    render status: :ok, json: { users: users }
-  end
 
   def create
     @user = User.new(user_params)
@@ -24,6 +18,8 @@ class UsersController < ApplicationController
 
   def user_params
     params.require(:user)
-      .permit(:first_name, :last_name, :email, :password, :password_confirmation)
+      .permit(
+        :first_name, :last_name, :email, :password, :password_confirmation
+      )
   end
 end
