--- app/views/layouts/mailer.text.erb
@@ -1,1 +1,0 @@
-<%= yield %>
