+++ config/locales/en.yml
@@ -30,9 +30,14 @@
 # available at https://guides.rubyonrails.org/i18n.html.
 
 en:
+  response:
+    success_message: "Thanks for voting!"
+    duplicate_error_message: "You have already voted!"
   authorization:
-    denied: "Access denied. You are not authorized to perform this action."
+    denied: "Access denied. You are not authorized to perform this action!"
   session:
-    could_not_auth: "Could not authenticate with the provided credentials."
-    incorrect_credentials: "Incorrect credentials, try again."
+    could_not_auth: "Could not authenticate with the provided credentials!"
+    incorrect_credentials: "Incorrect credentials, try again!"
   successfully_created: "%{entity} was successfully created!"
+  successfully_updated: "%{entity} was successfully updated!"
+  successfully_deleted: "%{entity} was successfully deleted!"
